+++ lib/open_api/generator.rb
@@ -1,7 +1,10 @@
+require 'active_support/hash_with_indifferent_access'
 require 'open_api/config'
 
 module OpenApi
   module Generator
+    extend self
+
     def self.included(base)
       base.extend ClassMethods
     end
@@ -18,31 +21,33 @@ module OpenApi
           # :nocov:
         end
         Dir[*Array(Config.doc_location)].each { |file| require file }
-        (doc_name || Config.docs.keys).map { |name| { name => generate_doc(name) } }.reduce({ }, :merge)
+        (doc_name || Config.docs.keys).map { |name| { name => generate_doc(name) } }.reduce({ }, :merge!)
       end
 
       def generate_doc(doc_name)
         settings = Config.docs[doc_name]
-        doc = { openapi: '3.0.0' }.merge(settings.slice :info, :servers).merge(
+        doc = { openapi: '3.0.0', **settings.slice(:info, :servers) }.merge!(
                 security: settings[:global_security], tags: [ ], paths: { },
                 components: {
                     securitySchemes: { }, schemas: { }, parameters: { }, requestBodies: { }
-                }.merge(settings[:components] || { })
+                }.merge!(settings[:components] || { })
               )
 
-        settings[:root_controller].descendants.each do |ctrl|
-          ctrl_infos = ctrl.instance_variable_get('@_ctrl_infos')
-          next if ctrl_infos.nil?
-          doc[:paths].merge! ctrl.instance_variable_get('@_api_infos') || { }
-          doc[:tags] << ctrl_infos[:tag]
-          doc[:components].merge!(ctrl_infos[:components] || { }) { |_key, l, r| l.merge(r) }
-          OpenApi.paths_index[ctrl.instance_variable_get('@_ctrl_path')] = doc_name
+        settings[:base_doc_class].descendants.each do |ctrl|
+          doc_info = ctrl.instance_variable_get('@doc_info')
+          next if doc_info.nil?
+
+          doc[:paths].merge!(ctrl.instance_variable_get('@api_info') || { })
+          doc[:tags] << doc_info[:tag]
+          doc[:components].deep_merge!(doc_info[:components] || { })
+          OpenApi.routes_index[ctrl.instance_variable_get('@route_base')] = doc_name
         end
+
         doc[:components].delete_if { |_, v| v.blank? }
         doc[:tags]  = doc[:tags].sort { |a, b| a[:name] <=> b[:name] }
         doc[:paths] = doc[:paths].sort.to_h
 
-        OpenApi.docs[doc_name] = ActiveSupport::HashWithIndifferentAccess.new(doc.delete_if { |_, v| v.blank? })
+        OpenApi.docs[doc_name] = HashWithIndifferentAccess.new(doc.delete_if { |_, v| v.blank? })
       end
 
       def write_docs(generate_files: true)
@@ -52,30 +57,31 @@ module OpenApi
         output_path = Config.file_output_path
         FileUtils.mkdir_p output_path
         max_length = docs.keys.map(&:size).sort.last
-        # puts '[ZRO] * * * * * *'
         docs.each do |doc_name, doc|
           puts "[ZRO] `#{doc_name.to_s.rjust(max_length)}.json` has been generated."
           File.open("#{output_path}/#{doc_name}.json", 'w') { |file| file.write JSON.pretty_generate doc }
         end
         # :nocov:
       end
+    end # end of module
+
+    def routes
+      @routes ||=
+          if (file = Config.rails_routes_file)
+            File.read(file)
+          else
+            # :nocov:
+            # ref https://github.com/rails/rails/blob/master/railties/lib/rails/tasks/routes.rake
+            require './config/routes'
+            all_routes = Rails.application.routes.routes
+            require 'action_dispatch/routing/inspector'
+            inspector = ActionDispatch::Routing::RoutesInspector.new(all_routes)
+            inspector.format(ActionDispatch::Routing::ConsoleFormatter.new, nil)
+            # :nocov:
+          end
     end
 
-    def self.routes_list
-      routes =
-        if (f = Config.rails_routes_file)
-          File.read(f)
-        else
-          # :nocov:
-          # ref https://github.com/rails/rails/blob/master/railties/lib/rails/tasks/routes.rake
-          require './config/routes'
-          all_routes = Rails.application.routes.routes
-          require 'action_dispatch/routing/inspector'
-          inspector = ActionDispatch::Routing::RoutesInspector.new(all_routes)
-          inspector.format(ActionDispatch::Routing::ConsoleFormatter.new, nil)
-          # :nocov:
-        end
-
+    def routes_list
       @routes_list ||= routes.split("\n").drop(1).map do |line|
         next unless line.match?('#')
         infos = line.match(/[A-Z|].*/).to_s.split(' ') # => [GET, /api/v1/examples/:id, api/v1/examples#index]
@@ -90,14 +96,14 @@ module OpenApi
       end.compact.group_by { |api| api[:action_path].split('#').first } # => { "api/v1/examples" => [..] }, group by paths
     end
 
-    def self.get_actions_by_ctrl_path(ctrl_path)
-      routes_list[ctrl_path]&.map do |action_info|
+    def get_actions_by_route_base(route_base)
+      routes_list[route_base]&.map do |action_info|
         action_info[:action_path].split('#').last
       end
     end
 
-    def self.find_path_httpverb_by(ctrl_path, action)
-      routes_list[ctrl_path]&.map do |action_info|
+    def find_path_httpverb_by(route_base, action)
+      routes_list[route_base]&.map do |action_info|
         if action_info[:action_path].split('#').last == action.to_s
           return [ action_info[:path], action_info[:http_verb].split('|').first ]
         end