+++ spec/dsl_spec.rb
@@ -2,21 +2,21 @@ require 'spec_helper'
 require 'dssl_helper'
 
 RSpec.describe OpenApi::DSL do
-  desc :ctrl_path, subject: :paths do
-    before_do { ctrl_path 'examples' }
-    make -> { api :action }, 'should not be mapped to goods#action', _it { be_nil }
+  desc :route_base, subject: :paths do
+    before_do { route_base 'examples' }
+    make -> { api :action }, 'is not mapped to goods#action', _it { be_nil }
 
-    make -> { api :index }, 'should be mapped to examples#index', has_key!: :'examples/index'
+    make -> { api :index }, 'is mapped to examples#index', has_key!: :'examples/index'
     focus_on :'examples/index', :get, :tags, 0
-    expect_it eq: 'Examples', desc: 'should be a Examples api'
+    expect_it eq: 'Examples', desc: 'is a Examples api'
 
-    after_do { ctrl_path 'goods' }
+    after_do { route_base 'goods' }
   end
 
 
-  desc :apis_tag do
+  desc :doc_tag do
     make -> do
-      apis_tag name: :Other, desc: 'tag desc', external_doc_url: 'url'
+      doc_tag name: :Other, desc: 'tag desc', external_doc_url: 'url'
       api :action
     end, has_keys!: %i[ tags paths ]
     focus_on :tags, 0
@@ -31,17 +31,17 @@ RSpec.describe OpenApi::DSL do
 
   desc :api, subject: :paths do
     context 'when this action is not configured routing' do
-      make -> { api :no_routing_action }, 'should refuse to be generated', _it { be_nil }
+      make -> { api :no_routing_action }, 'refuses to be generated', _it { be_nil }
     end
 
     context 'when this action can be accessed through multiple HTTP methods (set through `match`, like `GET|POST`)' do
-      make -> { api :change_onsale }, 'should match and generate both HTTP methods',
+      make -> { api :change_onsale }, 'matches and generate both HTTP methods',
            has_keys: { 'goods/{id}/change_onsale': %i[ post patch ] }
     end
 
     context 'when this action can be accessed through multiple HTTP methods (not set through `match`)' do
-      make -> { api :change_onsale }, 'should match the first HTTP method',
-           has_keys: { 'goods/{id}/change_onsale': [:patch] }
+      make -> { api :update }, 'matches and generate both HTTP methods',
+           has_keys: { 'goods/{id}': %i[ put patch ] }
     end
   end
 
@@ -52,7 +52,7 @@ RSpec.describe OpenApi::DSL do
         api_dry { resp :success, 'success response' }
         api :create
         api :index
-      end, 'should make all actions have a :success response',
+      end, 'makes all actions have a :success response',
            has_keys: { goods: [ get: [responses: [:success]], post: [responses: [:success]] ] }
     end