+++ spec/support/api_doc.rb
@@ -3,14 +3,14 @@ class ApiDoc < Object
 
   class << self
     def undo_dry
-      @_api_dry_blocks = nil
+      @zro_dry_blocks = nil
     end
 
     def inherited(subclass)
       super
       subclass.class_eval do
         break unless self.name.match?(/sController|sDoc/)
-        ctrl_path self.name.sub('Doc', '').downcase.gsub('::', '/') if self.name.match?(/sDoc/)
+        route_base self.name.sub('Doc', '').downcase.gsub('::', '/') if self.name.match?(/sDoc/)
       end
     end
   end